fix disabling dark mode
Previously, when disabling dark mode with `->darkMode(false)`, the design is to use the `system` default theme.

It will be a hassle if developers want to force the users to use light mode. because now the theme will be dependant to user's default theme

This PR will enable filament to force users to use light mode by simply use `->forceLightMode()`